% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterr.R
\name{filterr}
\alias{filterr}
\title{filterr
Doesn't do anything yet}
\usage{
filterr(df, var, test)
}
\arguments{
\item{df}{a dataframe}

\item{var}{a variable}

\item{test}{a test}
}
\value{
something some day
}
\description{
filterr
Doesn't do anything yet
}
