% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiniDplyR-package.R
\docType{package}
\name{MiniDplyR-package}
\alias{MiniDplyR}
\alias{MiniDplyR-package}
\title{MiniDplyR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Bjarke Hviid-Vyff \email{bhcor@dadlnet.dk}

}
\keyword{internal}
