% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectr.R
\name{selectr}
\alias{selectr}
\title{selectr}
\usage{
selectr(df, var)
}
\arguments{
\item{df}{a dataframe}

\item{x}{a vector of numeric indice(s) or character string(s)}
}
\value{
returns a data frame with a subset of columns
}
\description{
selects a subset of columns
}
\examples{
# returns the first column of mtcars
selectr(mtcars,1)

# returns the first column of mtcars
selectr(mtcars,"mpg")


# returns the 2nd and 3rd column of mtcars
selectr(mtcars,c(2,3)) 

}
